requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0 